+++ src/elements/response.js
@@ -1,3 +1,4 @@
+const Action = require('./action');
 const defaults = require('lodash.defaults');
 const NamedElement = require('./named-element');
 const TransactionMixin = require('./transaction-mixin');
@@ -33,5 +34,28 @@ module.exports = class Response extends TransactionMixin(NamedElement) {
     id += this.name || `Response ${this.statusCode}`;
 
     return this.getUniqueId(id);
+  }
+
+  get bodySchema() {
+    let property = 'responseBodySchema';
+
+    if (this.statusCode >= 300) {
+      property = 'responseErrorSchema';
+    }
+
+    return this._bodySchema || this.ancestor(Action)[property];
+  }
+
+  set bodySchema(schema) {
+    this._bodySchema = schema;
+  }
+
+  get bodySchemaSourcemap() {
+    return this._bodySchemaSourcemap ||
+      this.ancestor(Action).responseBodySchemaSourcemap;
+  }
+
+  set bodySchemaSourcemap(sourcemap) {
+    this._bodySchemaSourcemap = sourcemap;
   }
 };